Do not convert missing value to eltype of CFDiskArray (#30)
* Do not convert missing value to eltype of CFDiskArray

According to the CF Conventions https://cfconventions.org/Data/cf-conventions/cf-conventions-1.12/cf-conventions.html#missing-data the missing data check should be done on the raw data before applying offset and scale_factor. Therefore it should also be in the element type of the raw data.

* Add test for missing value check

The missing_value would be missed, if we convert it to the element type because of Floating point inaccuracies.

---------

Co-authored-by: Felix Cremer <felix.cremer@bgc-jena.mpg.de>